Sunrise/Set, Upgrades, Wave System
Another decent day, but should have shut it down an hour or two ago.  Took good breaks today and got a decent amount done.
- Fixed sunrise/sunset bug when skipping the day cycle during transition and locked it to camera
- Ended up refactoring the upgrades and upgrade system to allow for easier additions.  This was a decent surgery with quick recovery time, lol.  I broke down the one UpgradeType enum into 4 separate ones, and created an enum from UpgradeFamily, which then contains specific types.  This took a bit of time but went very smooth.
- Some minor adjustments to upgrades UI
- Addes message when building with not enough Credits
- Wave system was not working correctly, found an issue with current wave initialization. Fixed.
- Changed base wave multiplier to 10 to allow for smoother increase in numbers over time
- Fixed up the wave system a bit and tried tuning a handful of waves into a smooth experience.  Turns out my design is still off a bit, so I spent some time experimenting and thinking up a better way.  I’ll be changing the constant waves to actually be constant until the sub waves end.  I may try a randomized list approach to wave generation.  Creating them by hand does not allow randomization the way I have it now.  So, by creating simple templates as wave bases, I can then have a random selection of these waveBases and also for constant waves, which could start a bit later.  Over time the list could grow.  Might use ‘boss’ waves to have added difficulty with predictability, and can add new enemies after bosses.
- Lots to try out and figure out tomorrow.  Id like to have more control over the flow of the waves.

+++ Mutant Mayhem/Assets/_Scripts/Waves/WaveBases/WaveSOBase.cs
@@ -9,13 +9,13 @@ public class WaveSOBase : ScriptableObject
     //[Header("Sub Waves")]
     public List<SubWaveSO> subWaves;
     public List<SubWaveStyleSO> subWaveStyles;
-    public List<int> subWaveMultipliers;
+    public List<float> subWaveMultipliers;
     public List<int> timesToTriggerSubWaves;
 
     //[Header("Constant Waves")]
     public List<SubWaveSO> constantWaves;
     public List<SubWaveStyleSO> constantWaveStyles;
-    public List<int> constantWaveMultipliers;
+    public List<float> constantWaveMultipliers;
     public List<int> timesToTriggerConstantWaves;
 
     //[Header("Other Variables")]
